use pytest-mock for the VariantFetcherProxy
